refactor(utils): extract NCCL helpers to dedicated module
Move NCCL library discovery functions to vllm/utils/nccl.py:
- find_nccl_library: NCCL/RCCL library resolver
- find_nccl_include_paths: NCCL header path resolver

This reduces vllm/utils/__init__.py complexity and improves
modularity for distributed computing utilities.

Contributes to #26900
